+++ lib/widgets/numeric_fields/card_settings_int.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
 import '../../card_settings.dart';
-import '../../interfaces/common_field_attributes.dart';
+import '../../interfaces/common_field_properties.dart';
 
 /// This is a password field. It obscures the entered text.
 class CardSettingsInt extends CardSettingsText