+++ lib/widgets/picker_fields/card_settings_color_picker.dart
@@ -9,7 +9,7 @@ import 'package:flutter_cupertino_settings/flutter_cupertino_settings.dart';
 import 'package:flutter_material_pickers/flutter_material_pickers.dart';
 
 import '../../card_settings.dart';
-import '../../interfaces/common_field_attributes.dart';
+import '../../interfaces/common_field_properties.dart';
 
 enum CardSettingsColorPickerType { colors, material, block }
 
@@ -169,8 +169,14 @@ class _CardSettingsColorPickerState extends FormFieldState<Color> {
                       CardSettings.of(context).labelWidth ??
                       120.0,
                   child: widget?.requiredIndicator != null
-                      ? Text((widget?.label ?? "") + ' *', style: ls,)
-                      : Text(widget?.label, style: ls,),
+                      ? Text(
+                          (widget?.label ?? "") + ' *',
+                          style: ls,
+                        )
+                      : Text(
+                          widget?.label,
+                          style: ls,
+                        ),
                 ),
                 contentWidget: Container(
                   height: 20.0,