+++ lib/widgets/picker_fields/card_settings_datetime_picker.dart
@@ -11,7 +11,7 @@ import 'package:flutter_cupertino_settings/flutter_cupertino_settings.dart';
 import 'package:intl/intl.dart';
 
 import '../../card_settings.dart';
-import '../../interfaces/common_field_attributes.dart';
+import '../../interfaces/common_field_properties.dart';
 
 /// This is the date picker field
 class CardSettingsDateTimePicker extends FormField<DateTime>
@@ -221,8 +221,14 @@ class _CardSettingsDateTimePickerState extends FormFieldState<DateTime> {
                       CardSettings.of(context).labelWidth ??
                       120.0,
                   child: widget?.requiredIndicator != null
-                      ? Text((widget?.label ?? "") + ' *', style: ls,)
-                      : Text(widget?.label, style: ls,),
+                      ? Text(
+                          (widget?.label ?? "") + ' *',
+                          style: ls,
+                        )
+                      : Text(
+                          widget?.label,
+                          style: ls,
+                        ),
                 ),
                 contentWidget: Text(
                   value == null ? '' : DateFormat.yMd().add_jms().format(value),