+++ lib/widgets/text_fields/card_settings_paragraph.dart
@@ -5,8 +5,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
 import '../../card_settings.dart';
-import '../../interfaces/common_field_attributes.dart';
-import '../../interfaces/text_field_attributes.dart';
+import '../../interfaces/common_field_properties.dart';
+import '../../interfaces/text_field_properties.dart';
 
 /// This is a password field. It obscures the entered text.
 class CardSettingsParagraph extends CardSettingsText