+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Releases
 
+## [1.9.10] - TBD
+
+- created a `divider` property on CardSettings and CardSettingsSections
+- removed the divider appearing from the last item in a section
+
 ## [1.9.9+hotfix.1] - 2020-07-01
 
 - fixed content alignment issue